requirements.txt
firebase_admin
Pyrebase4
streamlit_option_menu
validate_email_address
#chromedrivermanager
streamlit
#webdriver-manager
#pywin32
selenium
openpyxl
setuptools
